Creating separate steps store
