+++ client/store/index.js
@@ -5,11 +5,13 @@ import {composeWithDevTools} from 'redux-devtools-extension'
 import user from './user'
 import category from './category'
 import {pathReducer} from './pathReducer'
+import step from './step'
 
 const reducer = combineReducers({
   user,
   pathReducer,
-  category
+  category,
+  step
 })
 const middleware = composeWithDevTools(
   applyMiddleware(thunkMiddleware, createLogger({collapsed: true}))
@@ -20,3 +22,4 @@ export default store
 export * from './user'
 export * from './category'
 export * from './pathReducer'
+export * from './step'
