+++ client/store/pathReducer.js
@@ -5,7 +5,6 @@ import axios from 'axios'
  */
 const GET_SINGLE_PATH = 'GET_SINGLE_PATH'
 const GET_PATHS_SINGLE_USER = 'GET_PATHS_SINGLE_USER'
-const GET_STEP_COMPLETIONS_FOR_USER = 'GET_STEP_COMPLETIONS_FOR_USER'
 const ADD_NEW_PATH = 'ADD_NEW_PATH'
 const DELETE_SINGLE_PATH = 'DELETE_SINGLE_PATH'
 
@@ -37,17 +36,17 @@ const getSinglePath = (path) => {
   }
 }
 
-const getStepCompletionSingleUser = (completed) => {
-  return {
-    type: GET_STEP_COMPLETIONS_FOR_USER,
-    completed
-  }
-}
-
 const deleteSinglePath = (name) => {
   return {
     type: DELETE_SINGLE_PATH,
     name
+  }
+}
+
+const toggleStepCompletion = (step) => {
+  return {
+    type: TOGGLE_STEP_COMPLETION,
+    step
   }
 }
 
@@ -96,13 +95,6 @@ export const getSinglePathThunk = (name) => {
   }
 }
 
-export const getStepCompletionSingleUserThunk = (pathName, username) => {
-  return async (dispatch) => {
-    const { data } = await axios.get(`/api/paths/${pathName}/user/${username}/completed`)
-    dispatch(getStepCompletionSingleUser(data))
-  }
-}
-
 export const deleteSinglePathThunk = (name) => {
   return async (dispatch) => {
     await axios.delete(`/api/paths/${name}`)
@@ -134,7 +126,6 @@ export const searchPathsInCategory = (categoryId, searchVal) => {
 const initialState = {
   allUserPaths: [],
   singlePath: [],
-  completedSteps: [],
   popularPathsInCategory: [],
   allPathsInCategory: [],
   searchedPathsInCategory: []
@@ -155,8 +146,6 @@ export const pathReducer = ( state = initialState, action) => { // eslint-disabl
       return {...state, allUserPaths: action.paths}
     case GET_SINGLE_PATH:
       return {...state, singlePath: action.path}
-    case GET_STEP_COMPLETIONS_FOR_USER:
-      return {...state, completedSteps: action.completed}
     case SET_POPULAR_PATHS_IN_CATEGORY:
       return {...state, popularPathsInCategory: action.paths}
     case SET_ALL_PATHS_IN_CATEGORY:
